% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{immigration}
\alias{immigration}
\title{Data on the Effect of Media Framing on Support for Immigration}
\format{An object of class \code{data.frame} with 265 rows and 12 columns.}
\usage{
immigration
}
\description{
A dataset containing 12 columns (including treatment, mediator, outcome, and
baseline and post-treatment confounders) and 265 individuals.
}
\references{
Brader, T., N. A. Valentino, and E. Suhay. 2008 "What Triggers Public Opposition to Immigration?
  Anxiety, Group Cues, and Immigration Threat." American Journal of Political Science 52(4): 959-978.
}
\keyword{datasets}
