% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{peace}
\alias{peace}
\title{Data on the Effect of Shared Democracy on Support for War}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1273 rows and 19 columns.}
\usage{
peace
}
\description{
A dataset containing 19 columns (including treatment, mediator, outcome, and
baseline and post-treatment confounders) and 1273 individuals.
}
\references{
Tomz, Michael R., and Jessica LP Weeks. 2013. "Public Opinion and the Democratic Peace."
  American Political Science Review 107(4): 849-865.
}
\keyword{datasets}
